move UI components into the UI folder
