+++ src/components/export-modal/export-modal.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import Modal from '../modal';
-import Button from '../button';
+import Modal from '../ui/modal';
+import Button from '../ui/button';
 import { toggleExportModal } from '../../actions';
 import exportGraph from './export-graph';
 import './export-modal.css';
