+++ src/components/ui/button/button.scss
@@ -1,7 +1,7 @@
 /** Imports **/
 
-@import '../../styles/variables';
-@import '../../styles/mixins';
+@import '../../../styles/variables';
+@import '../../../styles/mixins';
 
 /** Variables **/
