+++ src/components/ui/dropdown/dropdown.scss
@@ -1,5 +1,5 @@
-@import '../../styles/extends';
-@import '../../styles/variables';
+@import '../../../styles/extends';
+@import '../../../styles/variables';
 
 /** Variables */
