+++ src/components/ui/icon-button/icon-button.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables';
-@import '../../styles/extends';
+@import '../../../styles/variables';
+@import '../../../styles/extends';
 
 .pipeline-icon-toolbar__button {
   position: relative;
