+++ src/components/ui/menu-option/menu-option.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 /** Variables */
