+++ src/components/ui/modal/modal.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 /** Variables **/
