+++ src/components/ui/switch/switch.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .switch {
   align-items: center;
