+++ src/components/ui/toggle/toggle.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .kui-theme--light {
   --color-toggle-on: #{$color-primary-2};
