+++ src/components/ui/tooltip/tooltip.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables';
-@import '../../styles/extends';
+@import '../../../styles/variables';
+@import '../../../styles/extends';
 
 $x-offset: 20px;
 $y-offset: 10px;
